requirements.txt
python-bitcoinrpc >= 0.1

